Add support for more advanced TLS
Unless the later version is specified, it apparently defaults to TLSv1, which is no longer considered secure, even if the JVM is run with TLSv1.1 or TLSv1.2 enabled. Users will still need to enable TLSv1.2 (or TLSv1.1) explicitly for the JVM running Jenkins for this functionality to take effect, at least for Java7.