Fix client marshaling problem
For some reason, Jenkins was having problems when trying to reload the `CodeDxClient` instance from the config.xml file for the job. The problem was related to being unable to instantiate an instance of `LayeredSocketConnectionFactory` (which is an interface). Why it was trying to do that rather than instantiating an implementing class is unknown.

Making the `client` property transient, and then reloading it only when it's needed solves the problem, as there's nothing in the client itself that needs to be persisted.