restore backwards compatibility for pre-2.4.0 codedx versions
This affects the behavior of the "only consider new" option, and the
analysis name handling. Pre-2.4.0 does not support analysis names, so
the publish action will no-op for older versions of Code Dx. The
"New" triage status was removed in 2.4.0, but can still be used as
it was in older versions of this plugin up until then.

The new CodeDxVersion class is responsible for understanding and
comparing version numbers, and knowing at which version certain features
were added or removed.