Make Archiver’s FileFilter implementation be serializable
Previously, the anonymous inner class used as the FileFilter for creating an archive of the workspace didn’t implement serializable. As a result, when the plugin was used in a distributed environment, exceptions were thrown. By making it a static inner class, we can implement serializable, and that should resolve the issue.